Added mail server functions
And new demo front-end
