+++ test1/app/index.js
@@ -1,86 +1,56 @@
- 
-
-const ft = require('./ft')
-const URIlib = require('./URI')
-
-const cheerio =  require('cheerio')
+const express = require('express');
 const http = require('http');
 const https = require("https");
+const URIlib = require('./URI')
+
+const test = require('./testmail'); 
+
 const nodemailer = require('nodemailer');
 const later = require('later');
 
- 
-const express = require('express');
-const bodyParser= require('body-parser')
- 
-
 const app = express();
 
-const MongoClient = require('mongodb').MongoClient
- 
-var mongo_url="mongodb://root:1234@ds159517.mlab.com:59517/test_database";
-
-var db
-
-
-MongoClient.connect(mongo_url, (err, database) => {
-  if (err) return console.log(err)
-  db = database
-  app.listen(3000, () => {
-    console.log('listening on 3000')
-  })
+app.listen(3000, () => {
+console.log('listening on 3000') 
 })
 
-
-
-app.use(bodyParser.urlencoded({extended: true}))
-
-app.set('view engine', 'ejs') 
-
-
-
  
-  
-
 app.get('/', (req, res) => {
-  
-  db.collection('quotes').find().toArray((err, result) => {
-    if (err) return console.log(err)
-    // renders index.ejs
-    res.render('index.ejs', {quotes: result}) 
-  })
+   res.render('index.ejs', {}) 
+})
+   
+
+app.get('/showmail', (req, res) => {
+
+ //   console.log(test.test_data.results_json.search_results);
+   res.render('mail.ejs', {json_data:test.test_data}) 
 })
 
 
-app.post('/quotes', (req, res) => {
- 
- //req.body
-//save_to_db={website:ft.website,input_json:ft.input_json}
-  db.collection('quotes').save(req.body, (err, result) => {
-    if (err) return console.log(err)
-
-    console.log('saved to database')
-    res.redirect('/')
-  })
-})
- 
 
 
-
-
-app.get('/mail', (req, res) => {
+app.get('/sendmail', (req, res) => {
 
  
-  send_mail();
-  console.log('mail sent')
+res.render('mail.ejs', {json_data:test.test_data}, function(err, html){
+    if (err == null) {
+      send_mail(html);
+      res.send({ message: "ok" });
+    } else {
+        res.send({ error: err });
+    }
+  });
+ 
+});
 
 
-  res.send({ message: "sent" });
-}) 
+ 
+
+ 
 
 
 
-function send_mail() {
+function send_mail(html) {
 
  var smtpConfig = {
     host: 'smtp.yandex.ru',
@@ -95,9 +65,9 @@ function send_mail() {
   var mailData = {
     from: 'mydiskmydisk@yandex.ru',
     to: 'amantels@gmail.com',
-    subject: 'Message title',
-    text: 'Plaintext version of the message',
-    html: 'HTML version of the message'
+    subject: 'FUTUBNB',
+    text: 'Only HTML here, sorry',
+    html: html
   };
 
   transporter.sendMail(mailData);
@@ -105,89 +75,4 @@ function send_mail() {
   console.log('mail sent')
 
 };
-
-
-
-app.post("/fetch", function (req, res, next) {
-
-  if (req.body) {
-
-    if (req.body.url === undefined) {
-      res.send({ message: "url cannot be undefined" });
-    }
-    if (req.body.preference === undefined) {
-      res.send({ message: "preference cannot be undefined" });
-    }    
-
-/*
-    var urlPrefix = req.body.url.match(/.*?:\/\//g);
     
-*/
-
-
-    var url_instance = new URIlib.URI(req.body.url);
-    var transport = (url_instance.getScheme() || "").toLowerCase() === "https"? https : http;
-
-    var httpParams = {
-      host: url_instance.getAuthority(),
-      headers: {'user-agent': 'Mozilla/5.0'},
-      path: (url_instance.getPath() || "" )+"?"+(url_instance.getQuery() || "")
-    }
-
-   // console.log(url_instance.getPath());
-//res.send(httpParams);
-    
-  
-    var transpot_info=transport.get(httpParams, function (result) {
-        processResponse(result);
-      }).on('error', function (e) {
-        res.send({ message: e.message });
-      });
-
-        console.log(transpot_info);
-
-   
-    /*if (urlPrefix !== undefined && urlPrefix !== null && urlPrefix[0] === "https://") {
-      var info=  https.get(req.body.url, function (result) {
-        processResponse(result);
-      }).on('error', function (e) {
-        res.send({ message: e.message });
-      });
-    //  console.log(info);
-    } else {
-      http.get(req.body.url, function (result) {
-        processResponse(result);
-      }).on('error', function (e) {
-        res.send({ message: e.message });
-      });
-    }*/
-
-    var processResponse = function (result) {
-      var data = "";
-      result.on("data", function (chunk) {
-        data += chunk;
-      });
-      result.on("end", function (chunk) {
-   //  res.send(data);
-   console.log(data);
-        $ = cheerio.load(data);
- 
-        search_json=JSON.parse(req.body.preference);
-        json_result=ft.check(search_json);
-        res.setHeader('Content-Type', 'application/json');
-        res.send(JSON.stringify(json_result));
- 
-      });
-    }
-
-  } 
-});
-
-
-//later.setInterval(send_mail, later.parse.text('every 20 secs'));
-/*
-span_test='<span class="priceContainer_1rxvror" data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0"><span class="text_1lobov3-o_O-size_regular_1n7tijc-o_O-weight_bold_1lymmfr-o_O-inline_10f0ge3" data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0.0"><span data-pricerate="true" data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0.0.1:1">61062</span><span data-symbolafter="true" data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0.0.1:2">₽</span></span><span class="text_1lobov3-o_O-size_small_h25x8d-o_O-weight_light_1jo1qff-o_O-inline_10f0ge3" data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0.1"><span data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0.1.1:1"> </span><span data-reactid=".2crad29vy80.0.1.0.3.2.1.0.$0.$grid_0.$0/=1$=01$15968794.$=1$15968794.0.2.0.2.0.0.0.1.1:3">/ месяц</span></span></span>';
-  $ = cheerio.load(span_test);
-
-  console.log($("span[class^='priceContainer_']").text());
-  */
