+++ test1/views/test.js
@@ -1,0 +1,2 @@
+
+test_data={"results_json":{"search_results":[{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":9037122,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.79498490531362,"lng":37.60886347558464,"name":"Домик для хороших людей в Москве!","person_capacity":1,"picture_count":30,"picture_url":"https://a2.muscache.com/im/pictures/96ed007c-bdd2-484a-85c0-30d59d4f456b.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/96ed007c-bdd2-484a-85c0-30d59d4f456b.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/baedaf18-473e-438a-9f53-f6a006e9d600.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/1ac377a8-3bc4-4706-b2af-bbe38981a7ba.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/560c0f72-3854-4cc4-be4f-468bd45d76b7.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/adb99ae3-9d0f-4dcb-a39b-eda45e1eeaa9.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Mar'ina Roshcha, Москва","reviews_count":5,"star_rating":4.0,"room_type":"Private room","user":{"first_name":"Ирина","id":9752998,"thumbnail_url":"https://a2.muscache.com/im/users/9752998/profile_pic/1385639571/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Ирина","id":9752998,"thumbnail_url":"https://a2.muscache.com/im/users/9752998/profile_pic/1385639571/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1022,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":5553415,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.79462387899849,"lng":37.60844253248523,"name":"Одноместные домики. Смотрите фото))","person_capacity":1,"picture_count":37,"picture_url":"https://a0.muscache.com/im/pictures/74711829/e2f743d0_original.jpg?aki_policy=x_medium","picture_urls":["https://a0.muscache.com/im/pictures/74711829/e2f743d0_original.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/74711862/1de08e8e_original.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/74924373/f01cc3e3_original.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/74711881/5f9f3ac4_original.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/74711915/0b1eb10f_original.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Mar'ina Roshcha, Moskva","reviews_count":15,"star_rating":4.5,"room_type":"Private room","user":{"first_name":"Ирина","id":9752998,"thumbnail_url":"https://a2.muscache.com/im/users/9752998/profile_pic/1385639571/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Ирина","id":9752998,"thumbnail_url":"https://a2.muscache.com/im/users/9752998/profile_pic/1385639571/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1005,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":15341017,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.76377831951252,"lng":37.65639738665447,"name":"Одноместный номер","person_capacity":1,"picture_count":1,"picture_url":"https://a2.muscache.com/im/pictures/ee26136c-698e-4564-88f8-89628ede01f2.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/ee26136c-698e-4564-88f8-89628ede01f2.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Basmanny, Moskva","reviews_count":1,"star_rating":0,"room_type":"Private room","user":{"first_name":"Хостел","id":13914736,"thumbnail_url":"https://a1.muscache.com/im/users/13914736/profile_pic/1396700229/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Хостел","id":13914736,"thumbnail_url":"https://a1.muscache.com/im/users/13914736/profile_pic/1396700229/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1030,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":1507192,"instant_bookable":true,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.58043636344341,"lng":37.66652423961152,"name":"Beautiful and cozy apartment outside city center","person_capacity":2,"picture_count":6,"picture_url":"https://a1.muscache.com/im/pictures/1a12c76a-98e1-4bf5-a9f1-69a51fc1d3e6.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/1a12c76a-98e1-4bf5-a9f1-69a51fc1d3e6.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/14b9137f-a1a7-41f7-8841-9546721f581a.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/353e8f9b-5a49-404b-8452-227c753e84e6.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/31e7777e-b806-4463-9991-8681ee850a65.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/df9476f4-2b73-4fa1-8332-f3ccdce38e7d.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Moskva","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Анна","id":8050866,"thumbnail_url":"https://a1.muscache.com/im/pictures/50b5ebf2-4bc9-4aa1-8433-08132d334014.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Анна","id":8050866,"thumbnail_url":"https://a1.muscache.com/im/pictures/50b5ebf2-4bc9-4aa1-8433-08132d334014.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":true,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1169,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":9262594,"instant_bookable":true,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.788062124572875,"lng":37.6143335952737,"name":"Кровать в общей комнате на 6 чел","person_capacity":2,"picture_count":3,"picture_url":"https://a1.muscache.com/im/pictures/58bdf3e5-e12f-4791-8771-d55859c98aa1.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/58bdf3e5-e12f-4791-8771-d55859c98aa1.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/0229c807-f0a6-48e5-b2ad-1ef8f9b1ffb2.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/c2580495-69d2-45bc-921d-bfd346aac55a.jpg?aki_policy=x_medium"],"property_type":"House","public_address":"Mar'ina Roshcha, Moskva","reviews_count":3,"star_rating":2.5,"room_type":"Private room","user":{"first_name":"Diana","id":7738943,"thumbnail_url":"https://a0.muscache.com/im/users/7738943/profile_pic/1375818785/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Diana","id":7738943,"thumbnail_url":"https://a0.muscache.com/im/users/7738943/profile_pic/1375818785/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":true,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1049,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":["en"],"id":11995536,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.767694239797535,"lng":37.411792817051804,"name":"My Room in Moscow City","person_capacity":3,"picture_count":11,"picture_url":"https://a2.muscache.com/im/pictures/e5c26695-811c-4980-bdae-7b633467008c.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/e5c26695-811c-4980-bdae-7b633467008c.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/54fe9ba4-4dd0-43a1-a927-9a8588a67a24.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/783a67fd-534f-42bf-8734-bc845dcfd968.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/f03e9342-1002-4eba-b021-bcf4a3a63bde.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/536d7802-f725-490d-b748-8fc4cb75230b.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Krylatskoye, Moskva","reviews_count":5,"star_rating":3.0,"room_type":"Private room","user":{"first_name":"Elena","id":24848479,"thumbnail_url":"https://a1.muscache.com/im/users/24848479/profile_pic/1418641948/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Elena","id":24848479,"thumbnail_url":"https://a1.muscache.com/im/users/24848479/profile_pic/1418641948/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1061,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":2,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":15021531,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.83130894646081,"lng":37.352069712696085,"name":"Сдается комната в уютной квартире","person_capacity":2,"picture_count":3,"picture_url":"https://a2.muscache.com/im/pictures/097cfd92-2c43-490c-a864-cb441250fa9f.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/097cfd92-2c43-490c-a864-cb441250fa9f.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/caa3c588-5ef8-468a-af77-13defccea06a.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/03123423-d138-4bc3-9a18-b79067219228.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Moscow","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Тамара","id":94817549,"thumbnail_url":"https://a0.muscache.com/im/pictures/496fc0d7-0394-4eb5-aef7-3ab1a19eb158.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Тамара","id":94817549,"thumbnail_url":"https://a0.muscache.com/im/pictures/496fc0d7-0394-4eb5-aef7-3ab1a19eb158.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":979,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":15127056,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.82029832928426,"lng":37.798905861469365,"name":"Room near park with double bed","person_capacity":2,"picture_count":33,"picture_url":"https://a1.muscache.com/im/pictures/5d8b494f-bda7-4206-adfa-067f43ac11a2.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/5d8b494f-bda7-4206-adfa-067f43ac11a2.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/2802cfe5-bf85-4f13-90b1-5514a981ed98.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/3912f313-9981-450e-ad65-0e62d425db44.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/8c26f280-be2c-4e0d-a6ac-1344a3b09b1c.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/fa44e2c3-4446-410a-9391-ea391639a385.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Golyanovo, Москва","reviews_count":4,"star_rating":5.0,"room_type":"Private room","user":{"first_name":"Евгений","id":91249108,"thumbnail_url":"https://a0.muscache.com/im/pictures/8a241a55-9d34-44c7-83cf-fea7824d25d3.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Евгений","id":91249108,"thumbnail_url":"https://a0.muscache.com/im/pictures/8a241a55-9d34-44c7-83cf-fea7824d25d3.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":919,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":13822251,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.61362598665459,"lng":37.596298048350896,"name":"комната на Пражской","person_capacity":2,"picture_count":3,"picture_url":"https://a1.muscache.com/im/pictures/b1c33671-1f5c-4475-89ae-36a44cb2712c.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/b1c33671-1f5c-4475-89ae-36a44cb2712c.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/a4a7fbda-7efd-40ec-ac92-c40e0e9908d8.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/bb94ccf7-5686-425d-a8e7-b9291bf6dbf6.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Moskva","reviews_count":7,"star_rating":4.0,"room_type":"Private room","user":{"first_name":"Мария","id":81436147,"thumbnail_url":"https://a1.muscache.com/im/pictures/640e033e-127c-441b-9903-30c03482ef1f.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Мария","id":81436147,"thumbnail_url":"https://a1.muscache.com/im/pictures/640e033e-127c-441b-9903-30c03482ef1f.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1185,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":15887227,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.87779250859782,"lng":37.51779096765421,"name":"Уютная тихая комната недалеко от аэропорта + парк!","person_capacity":2,"picture_count":6,"picture_url":"https://a2.muscache.com/im/pictures/d69c0db6-f09c-4f99-88e9-0ce013ed9508.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/d69c0db6-f09c-4f99-88e9-0ce013ed9508.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/c8c5dbee-5134-4c16-bd5d-8dbddb65d17e.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/4ebb922e-7382-4408-8a08-dbc27ffa364b.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/892f3061-b661-4661-99ac-c8d574ee00b8.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/abc23ec6-cbea-4576-8251-a1deec9a7eae.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Moskva","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Mikhail","id":102896928,"thumbnail_url":"https://a0.muscache.com/im/pictures/6d4c2231-717c-4d45-8754-dd2363e52574.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Mikhail","id":102896928,"thumbnail_url":"https://a0.muscache.com/im/pictures/6d4c2231-717c-4d45-8754-dd2363e52574.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1236,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":16013610,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.85987954359261,"lng":37.42016370628792,"name":"Artist`s appartement near airport Sheremet`evo","person_capacity":2,"picture_count":17,"picture_url":"https://a1.muscache.com/im/pictures/92b7a631-14c0-47b4-98d1-c589cc9a3af3.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/92b7a631-14c0-47b4-98d1-c589cc9a3af3.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/11b20240-5b5e-4a32-8b6a-0e3eba30fcfb.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/adfe0820-caab-4830-b790-f05558f352a7.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/c7c9efe3-8ff1-4379-8621-d59ff42905fa.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/54f1a79f-673d-4c46-96bc-5225e7aa40cc.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Severnoye Tushino, Moskva","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Natalia","id":104124489,"thumbnail_url":"https://a1.muscache.com/im/pictures/3195be10-16bc-4463-9740-b9af7c0f5d0c.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Natalia","id":104124489,"thumbnail_url":"https://a1.muscache.com/im/pictures/3195be10-16bc-4463-9740-b9af7c0f5d0c.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1184,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":16,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":16044613,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.73607964217383,"lng":37.652150181912006,"name":"Театр на Таганке, Музей Высоцкого.","person_capacity":16,"picture_count":25,"picture_url":"https://a2.muscache.com/im/pictures/b074ffe0-c759-467b-93b3-4f108285c6ad.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/b074ffe0-c759-467b-93b3-4f108285c6ad.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/9e14c457-c975-4e9d-8331-078b9e572163.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/8b9de062-86e8-4fe5-a204-367339dd9167.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/1440033a-8a8f-4892-90d5-ab241e022643.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/062d9b33-4541-4b18-a590-60a7d0be127d.jpg?aki_policy=x_medium"],"property_type":"Hostel","public_address":"Taganskiy, Moskva","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Юрий","id":99050623,"thumbnail_url":"https://a1.muscache.com/im/pictures/854446d8-99cb-42e3-a2c2-bccdd7158ba5.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Юрий","id":99050623,"thumbnail_url":"https://a1.muscache.com/im/pictures/854446d8-99cb-42e3-a2c2-bccdd7158ba5.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":670,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":2,"airbnb_plus_enabled":false,"extra_host_languages":["de","en","he"],"id":3956283,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.76312504568851,"lng":37.64807738121276,"name":"Отдельная комната рядом с Кремлем!","person_capacity":3,"picture_count":23,"picture_url":"https://a0.muscache.com/im/pictures/49767473/2803ba6e_original.jpg?aki_policy=x_medium","picture_urls":["https://a0.muscache.com/im/pictures/49767473/2803ba6e_original.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/49767484/834a9e72_original.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/49767498/526a1212_original.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/49767514/e4b8f71d_original.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/49767531/96fcee3b_original.jpg?aki_policy=x_medium"],"property_type":"Dorm","public_address":"Basmanny, Москва","reviews_count":5,"star_rating":4.0,"room_type":"Private room","user":{"first_name":"Elena","id":12718222,"thumbnail_url":"https://a0.muscache.com/im/users/12718222/profile_pic/1393764416/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Elena","id":12718222,"thumbnail_url":"https://a0.muscache.com/im/users/12718222/profile_pic/1393764416/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1129,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":null,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":16062169,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.728857131743176,"lng":37.6653862818961,"name":"Уютная комната в 5 минутах от Кремля","person_capacity":2,"picture_count":4,"picture_url":"https://a1.muscache.com/im/pictures/26d9e59b-95cf-41b6-bf6d-f5f9fc5e8ac3.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/26d9e59b-95cf-41b6-bf6d-f5f9fc5e8ac3.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/00c49d4d-bf7c-4abc-8eac-36fecfb3de9c.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/6d81c696-99fa-45f0-9feb-059741e69b5c.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/4cb5b852-e206-41b3-810b-e03b2f51ffc8.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Yuzhnoportovyy, Moscow","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Семен","id":64534764,"thumbnail_url":"https://a2.muscache.com/im/pictures/5140fbe6-d011-499c-9ce7-30258cf8ef23.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Семен","id":64534764,"thumbnail_url":"https://a2.muscache.com/im/pictures/5140fbe6-d011-499c-9ce7-30258cf8ef23.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":991,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":["en","uk"],"id":12900178,"instant_bookable":true,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.71728129170619,"lng":37.82315491704453,"name":"Заселение круглосуточно","person_capacity":2,"picture_count":5,"picture_url":"https://a2.muscache.com/im/pictures/af128236-86f9-4cd3-b47e-f19ebbe9e515.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/af128236-86f9-4cd3-b47e-f19ebbe9e515.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/827290de-b040-405b-84cb-e2fc2e06cb90.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/a1763ee1-bdee-4a64-a099-bb245bd09414.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/48a6187c-420a-4c4e-a28b-a7ab5683e556.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/88f36b29-2295-4dc2-a9cc-4935b6b02cb7.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Veshnyaki, Москва","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Александр","id":70661437,"thumbnail_url":"https://a2.muscache.com/im/pictures/881c739d-ad7f-4290-9f07-67358a14e30a.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Александр","id":70661437,"thumbnail_url":"https://a2.muscache.com/im/pictures/881c739d-ad7f-4290-9f07-67358a14e30a.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":true,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1047,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":["en"],"id":4195774,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.75156594208701,"lng":37.79504720200374,"name":"Welcome to Moscow","person_capacity":2,"picture_count":6,"picture_url":"https://a0.muscache.com/im/pictures/82f4d224-956c-4cb0-a673-e7f9851245ab.jpg?aki_policy=x_medium","picture_urls":["https://a0.muscache.com/im/pictures/82f4d224-956c-4cb0-a673-e7f9851245ab.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/1edf6742-1c4a-4a6d-ad3e-31b81df343d6.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/1a930859-e536-42e4-a2e9-8c406826de13.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/52855559/a08a3c94_original.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/1219a06c-99fe-4278-bb2f-376d2c80ee88.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Novogireevo, Moscow","reviews_count":1,"star_rating":0,"room_type":"Private room","user":{"first_name":"Evgeny","id":21730244,"thumbnail_url":"https://a0.muscache.com/im/users/21730244/profile_pic/1411806261/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Evgeny","id":21730244,"thumbnail_url":"https://a0.muscache.com/im/users/21730244/profile_pic/1411806261/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1138,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":[],"id":15501391,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":true,"lat":55.707696410050346,"lng":37.67751817616471,"name":"COSY ROOM","person_capacity":1,"picture_count":5,"picture_url":"https://a2.muscache.com/im/pictures/3260b31a-e921-4492-be3a-1e7bec45b97e.jpg?aki_policy=x_medium","picture_urls":["https://a2.muscache.com/im/pictures/3260b31a-e921-4492-be3a-1e7bec45b97e.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/02532e4f-acdd-4abe-9aef-5a5c82186653.jpg?aki_policy=x_medium","https://a2.muscache.com/im/pictures/b72ddb09-fa7f-4c7a-8759-b75884e8ec12.jpg?aki_policy=x_medium","https://a0.muscache.com/im/pictures/62b7e348-347d-4339-abce-a534131846ec.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/30d5542c-48e7-4c2c-88d1-55ee216efb5d.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Yuzhnoportovyy, Москва","reviews_count":0,"star_rating":0,"room_type":"Private room","user":{"first_name":"Юлия","id":21083102,"thumbnail_url":"https://a0.muscache.com/im/users/21083102/profile_pic/1410190374/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Юлия","id":21083102,"thumbnail_url":"https://a0.muscache.com/im/users/21083102/profile_pic/1410190374/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1030,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null},{"listing":{"bedrooms":1,"beds":1,"airbnb_plus_enabled":false,"extra_host_languages":["en"],"id":14895278,"instant_bookable":false,"is_business_travel_ready":false,"is_new_listing":false,"lat":55.61851367712665,"lng":37.72545084471787,"name":"Чистая, уютная, домашняя атмосфера.","person_capacity":2,"picture_count":7,"picture_url":"https://a1.muscache.com/im/pictures/fda57956-a60f-4c05-af6d-f305921bcb75.jpg?aki_policy=x_medium","picture_urls":["https://a1.muscache.com/im/pictures/fda57956-a60f-4c05-af6d-f305921bcb75.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/289e760f-2b35-4cb2-98da-77d5f6579406.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/ad9f4c81-2dbb-45a0-aefe-5563ba720c9f.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/e6e34304-6be6-4a99-95a3-9a0d939982e7.jpg?aki_policy=x_medium","https://a1.muscache.com/im/pictures/5311ce86-be5e-4333-b20c-5cf0ec01b9d6.jpg?aki_policy=x_medium"],"property_type":"Apartment","public_address":"Москва","reviews_count":1,"star_rating":0,"room_type":"Private room","user":{"first_name":"Madlena","id":22521990,"thumbnail_url":"https://a0.muscache.com/im/users/22521990/profile_pic/1413654489/original.jpg?aki_policy=profile_medium","is_superhost":false},"primary_host":{"first_name":"Madlena","id":22521990,"thumbnail_url":"https://a0.muscache.com/im/users/22521990/profile_pic/1413654489/original.jpg?aki_policy=profile_medium","is_superhost":false},"coworker_hosted":false,"listing_tags":[]},"pricing_quote":{"available":true,"can_instant_book":false,"check_in":"2016-11-24","check_out":"2016-12-14","guests":1,"rate":{"amount":1125,"currency":"RUB"},"rate_type":"nightly","is_good_price":null,"average_booked_price":null},"viewed_at":null}],"metadata":{"guidebook":null,"search":{"business_travel_reward_data":{},"has_airbnb_plus_availability":false,"has_business_travel_ready_availability":false,"is_last_minute_eligible":false,"last_minute_show_dist_sort":false,"mobile_session_id":"h6QeJGZt","native_currency":"RUB","price_type":"nightly","price_range_max_native":50000,"price_range_min_native":640,"search_id":"38f92a5d-7d27-4c8f-b780-44379df0d177"},"pagination":{"next_offset":18,"result_count":18},"facets":{"bedrooms":[{"key":1,"value":"1","count":134},{"key":0,"value":"0","count":134}],"bathrooms":[{"key":5.0,"value":"5.0","count":1},{"key":4.5,"value":"4.5","count":1},{"key":4.0,"value":"4.0","count":3},{"key":3.5,"value":"3.5","count":3},{"key":3.0,"value":"3.0","count":7},{"key":2.5,"value":"2.5","count":7},{"key":2.0,"value":"2.0","count":15},{"key":1.5,"value":"1.5","count":23},{"key":1.0,"value":"1.0","count":133},{"key":0.5,"value":"0.5","count":133},{"key":0.0,"value":"0.0","count":134}],"beds":[{"key":16,"value":"16","count":1},{"key":15,"value":"15","count":1},{"key":14,"value":"14","count":1},{"key":13,"value":"13","count":1},{"key":12,"value":"12","count":3},{"key":11,"value":"11","count":3},{"key":10,"value":"10","count":5},{"key":9,"value":"9","count":5},{"key":8,"value":"8","count":7},{"key":7,"value":"7","count":7},{"key":6,"value":"6","count":8},{"key":5,"value":"5","count":8},{"key":4,"value":"4","count":13},{"key":3,"value":"3","count":21},{"key":2,"value":"2","count":41},{"key":1,"value":"1","count":133},{"key":0,"value":"0","count":134}],"availability":[{"key":"Instant Book","value":"Instant Book","count":6}],"room_type":[{"key":"Entire home/apt","value":"Entire home/apt","count":27},{"key":"Private room","value":"Private room","count":134},{"key":"Shared room","value":"Shared room","count":300}],"hosting_amenity_ids":[{"key":8,"value":"Kitchen","count":125},{"key":3,"value":"Internet","count":65},{"key":1,"value":"TV","count":84},{"key":40,"value":"Essentials","count":98},{"key":41,"value":"Shampoo","count":74},{"key":30,"value":"Heating","count":111},{"key":5,"value":"Air conditioning","count":26},{"key":33,"value":"Washer","count":116},{"key":34,"value":"Dryer","count":25},{"key":9,"value":"Free parking on premises","count":84},{"key":4,"value":"Wireless Internet","count":112},{"key":2,"value":"Cable TV","count":21},{"key":16,"value":"Breakfast","count":33},{"key":12,"value":"Pets allowed","count":26},{"key":31,"value":"Family/kid friendly","count":53},{"key":32,"value":"Suitable for events","count":12},{"key":11,"value":"Smoking allowed","count":39},{"key":6,"value":"Wheelchair accessible","count":6},{"key":21,"value":"Elevator in building","count":75},{"key":27,"value":"Indoor fireplace","count":7},{"key":28,"value":"Buzzer/wireless intercom","count":45},{"key":14,"value":"Doorman","count":13},{"key":7,"value":"Pool","count":3},{"key":25,"value":"Hot tub","count":12},{"key":15,"value":"Gym","count":8},{"key":43,"value":"24-hour check-in","count":40},{"key":44,"value":"Hangers","count":67},{"key":46,"value":"Iron","count":65},{"key":45,"value":"Hair dryer","count":58},{"key":47,"value":"Laptop friendly workspace","count":57},{"key":35,"value":"Smoke detector","count":31},{"key":36,"value":"Carbon monoxide detector","count":16},{"key":37,"value":"First aid kit","count":53},{"key":38,"value":"Safety card","count":16},{"key":39,"value":"Fire extinguisher","count":29},{"key":42,"value":"Lock on bedroom door","count":26}],"top_amenities":[{"key":4,"value":"Wireless Internet","count":112},{"key":7,"value":"Pool","count":3},{"key":8,"value":"Kitchen","count":125}],"neighborhood_facet":[{"key":"Akademicheskiy","value":"Akademicheskiy","count":2},{"key":"Basmanny","value":"Basmanny","count":4},{"key":"Biryulevo Vostochnoye","value":"Biryulevo Vostochnoye","count":1},{"key":"Bogorodskoye","value":"Bogorodskoye","count":1},{"key":"Brateyevo","value":"Brateyevo","count":1},{"key":"Butyrskiy","value":"Butyrskiy","count":1},{"key":"Chertanovo Tsentral'noye","value":"Chertanovo Tsentral'noye","count":1},{"key":"Danilovskiy","value":"Danilovskiy","count":1},{"key":"Dorogomilovo","value":"Dorogomilovo","count":2},{"key":"Golyanovo","value":"Golyanovo","count":1},{"key":"Kapotnya","value":"Kapotnya","count":1},{"key":"Kon'kovo","value":"Kon'kovo","count":1},{"key":"Koptevo","value":"Koptevo","count":1},{"key":"Krylatskoye","value":"Krylatskoye","count":2},{"key":"Levoberezhnyy","value":"Levoberezhnyy","count":1},{"key":"Lyublino","value":"Lyublino","count":1},{"key":"Mar'ina Roshcha","value":"Mar'ina Roshcha","count":3},{"key":"Mar'ino","value":"Mar'ino","count":2},{"key":"Marfino","value":"Marfino","count":1},{"key":"Meshchanskiy","value":"Meshchanskiy","count":3},{"key":"Novogireevo","value":"Novogireevo","count":1},{"key":"Orekhovo-Borisovo Severnoye","value":"Orekhovo-Borisovo Severnoye","count":1},{"key":"Ostankinskiy","value":"Ostankinskiy","count":1},{"key":"Otradnoye","value":"Otradnoye","count":4},{"key":"Perovo","value":"Perovo","count":1},{"key":"Presnenskiy","value":"Presnenskiy","count":2},{"key":"Severnoye Medvedkovo","value":"Severnoye Medvedkovo","count":2},{"key":"Severnoye Tushino","value":"Severnoye Tushino","count":2},{"key":"Shchukino","value":"Shchukino","count":1},{"key":"Sokol","value":"Sokol","count":1},{"key":"Taganskiy","value":"Taganskiy","count":1},{"key":"Troparevo-Nikulino","value":"Troparevo-Nikulino","count":1},{"key":"Tsaritsyno","value":"Tsaritsyno","count":1},{"key":"Veshnyaki","value":"Veshnyaki","count":2},{"key":"Vostochnoye Degunino","value":"Vostochnoye Degunino","count":1},{"key":"Voykovskiy","value":"Voykovskiy","count":2},{"key":"Vykhino-Zhulebino","value":"Vykhino-Zhulebino","count":1},{"key":"Yuzhnoportovyy","value":"Yuzhnoportovyy","count":2},{"key":"Zamoskvorechye","value":"Zamoskvorechye","count":1},{"key":"Zapadnoye Degunino","value":"Zapadnoye Degunino","count":1}],"top_neighborhoods":[{"key":"Otradnoye","value":"Otradnoye","count":4},{"key":"Basmanny","value":"Basmanny","count":4},{"key":"Mar'ina Roshcha","value":"Mar'ina Roshcha","count":3},{"key":"Meshchanskiy","value":"Meshchanskiy","count":3}],"property_type_id":[{"key":2,"value":"House","count":20},{"key":1,"value":"Apartment","count":100},{"key":3,"value":"Bed & Breakfast","count":7},{"key":45,"value":"Hostel","count":1},{"key":9,"value":"Dorm","count":3},{"key":36,"value":"Townhouse","count":1},{"key":33,"value":"Other","count":1}],"facilities_amenities":[{"key":21,"value":"Elevator in building","count":75},{"key":9,"value":"Free parking on premises","count":84},{"key":15,"value":"Gym","count":8},{"key":25,"value":"Hot tub","count":12},{"key":7,"value":"Pool","count":3},{"key":6,"value":"Wheelchair accessible","count":6}],"house_rules_amenities":[{"key":32,"value":"Suitable for events","count":12},{"key":12,"value":"Pets allowed","count":26},{"key":11,"value":"Smoking allowed","count":39}],"other_amenities":[{"key":8,"value":"Kitchen","count":125},{"key":3,"value":"Internet","count":65},{"key":1,"value":"TV","count":84},{"key":41,"value":"Shampoo","count":74},{"key":30,"value":"Heating","count":111},{"key":5,"value":"Air conditioning","count":26},{"key":33,"value":"Washer","count":116},{"key":34,"value":"Dryer","count":25},{"key":4,"value":"Wireless Internet","count":112},{"key":2,"value":"Cable TV","count":21},{"key":16,"value":"Breakfast","count":33},{"key":31,"value":"Family/kid friendly","count":53},{"key":27,"value":"Indoor fireplace","count":7},{"key":28,"value":"Buzzer/wireless intercom","count":45},{"key":14,"value":"Doorman","count":13},{"key":43,"value":"24-hour check-in","count":40},{"key":44,"value":"Hangers","count":67},{"key":46,"value":"Iron","count":65},{"key":45,"value":"Hair dryer","count":58},{"key":47,"value":"Laptop friendly workspace","count":57},{"key":42,"value":"Lock on bedroom door","count":26}],"top_other_amenities":[{"key":4,"value":"Wireless Internet","count":112},{"key":1,"value":"TV","count":84},{"key":8,"value":"Kitchen","count":125}],"languages":[{"key":4,"value":"Deutsch","count":3},{"key":1,"value":"English","count":65},{"key":64,"value":"Español","count":8},{"key":2,"value":"Français","count":9},{"key":16,"value":"Italiano","count":6},{"key":2097152,"value":"Polski","count":1},{"key":33554432,"value":"Svenska","count":1},{"key":2048,"value":"Türkçe","count":1},{"key":32,"value":"Русский","count":127},{"key":1073741824,"value":"українська","count":4},{"key":1048576,"value":"עברית","count":2},{"key":256,"value":"العربية","count":6},{"key":128,"value":"中文","count":1}],"first_languages":[{"key":1,"value":"English","count":65},{"key":64,"value":"Español","count":8},{"key":2,"value":"Français","count":9}],"description_languages":[{"key":4,"value":"Deutsch","count":1},{"key":1,"value":"English","count":40},{"key":2,"value":"Français","count":1},{"key":32,"value":"Русский","count":119}],"first_description_languages":[{"key":1,"value":"English","count":40},{"key":2,"value":"Français","count":1},{"key":4,"value":"Deutsch","count":1}]},"listing_tags":{},"listings_count":84,"urgency_commitment":{"message":{"headline":"Only 14% of listings are left for these dates.","body":"We recommend booking a place soon.","contextual_message":null},"message_type":"percent-available","show_percent_listings_left_message":true,"listings_left_as_percent":14},"market_info":{"headline":"Places in Moscow went fast last year","body":"Wondering when to book? This shows the percentage of available places for your travel dates at the same time last year. Places are likely to get booked at a similar rate this year.","data":[]},"price_histogram":{"average_price":2422,"histogram":[14,6,8,7,35,62,119,76,135,106,101,53,46,23,23,12,6,19,4,2,1,0,3,1,1,0,3,2,1,2,0,0,0,0,0,0,0,0,1,2,0,0,0,0,0,0,0,0,1,5]},"search_feed_items":[],"avg_price_by_room_type":{"ratio":{"Entire home/apt":0.32142857142857145,"Private room":1.5952380952380953,"Shared room":3.5714285714285716},"avg_price":{"Private room":14,"Shared room":9,"Entire home/apt":13}},"messages":{},"overrides":{},"breadcrumbs":[{"location_name":"Russia","canonical_location_name":"Russia","type":"country"},{"location_name":"Moscow","canonical_location_name":"Moscow, Russia","type":"state"},{"location_name":"Moscow","canonical_location_name":"Moscow, Russia","type":"city"}],"guided_search":{"avg_gbm_score":15.638719611697727,"avg_usd_price":16.019729276808953,"min_stars":4.0,"min_review_count":5.0},"location":{"canonical_location":"Moscow, Russia","display_location":"Moscow"},"geography":{"accuracy":4,"precision":"city","country_code":"RU","country":"Russia","result_type":"locality","state_short":"Moscow","state":"Moscow","city":"Moscow","lat":55.755826,"lng":37.6173,"place_id":"ChIJybDUc_xKtUYRTM9XV8zWRD0"},"refund_price":null}}}
